+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add locale to avoid console warnings. Commit [becf5e6] @peaceful-james
 
+### Fixed
+- Trigger `phx-change` event for input with `phx-debounce="blur"`
+
 ## [0.6.2] 2025-04-24
 ### Changed
 - `Frame.evaluate/3`: Don't transform map keys in return value. Example: `js {camelCase: {a: 1}} -> ex %{"camelCase": %{"a": 1}}`. Previously attempted to underscore and atom-ize keys, which led to issue [#24](https://github.com/ftes/phoenix_test_playwright/pull/24). Commit [5ff530]